Fixed bug where habits did not appear on initial sign in, and moved the username prompt to the create account screen
