+++ CMPUT301Project/app/src/main/java/com/example/GroupProject/MainActivity.java
@@ -58,16 +58,12 @@ public class MainActivity extends AppCompatActivity {
 
         FirebaseFirestore db;
 
-        if (!((GroupProject) this.getApplication()).isSignedIn()) {
-            Intent intent = new Intent(this, EmailPasswordActivity.class);
-            startActivity(intent);
-        }
+//        if (!((GroupProject) this.getApplication()).isSignedIn()) {
+//            Intent intent = new Intent(this, EmailPasswordActivity.class);
+//            startActivity(intent);
+//        }
 
         Intent intent = getIntent();
-
-        if (intent.getBooleanExtra("CREATED", false)) {
-            promptUsername();
-        }
 
         BottomNavigationView bottomNavigationView = findViewById(R.id.bottomNavigationBar);
         bottomNavigationView.setOnNavigationItemSelectedListener(item -> {
@@ -87,7 +83,7 @@ public class MainActivity extends AppCompatActivity {
             return true;
         });
 
-        int selected = intent.getIntExtra("SELECTED", 0);
+        int selected = intent.getIntExtra("SELECTED", R.id.habit);
         bottomNavigationView.setSelectedItemId(selected);
 
         // Access a Cloud Firestore instance from your Activity
@@ -101,40 +97,5 @@ public class MainActivity extends AppCompatActivity {
     public void addHabit(View view) {
         Intent intent = new Intent(this, AddHabitActivity.class);
         startActivity(intent);
-    }
-
-
-    private void promptUsername() {
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        builder.setTitle("Username Prompt");
-
-        final EditText input = new EditText(this);
-        builder.setView(input);
-        GroupProject thisGP = (GroupProject) this.getApplicationContext();
-
-        builder.setPositiveButton("Confirm", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                FirebaseFirestore db = FirebaseFirestore.getInstance();
-                Map<String, Object> docData = new HashMap<>();
-                docData.put("Joined", Timestamp.now());
-                docData.put("Email", thisGP.getEmail());
-                username = input.getText().toString();
-                docData.put("Username", username);
-                thisGP.setUsername(username);
-                db.collection("Users").document(username).set(docData, SetOptions.merge());
-            }
-        });
-        builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                dialog.cancel();
-                promptUsername();
-                Toast.makeText(MainActivity.this, "Username needs to be entered for account.",
-                        Toast.LENGTH_LONG).show();
-            }
-        });
-
-        builder.show();
     }
 }
