+++ CMPUT301Project/app/src/main/java/com/example/GroupProject/ProfileFragment.java
@@ -21,6 +21,7 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageButton;
+import android.widget.Toast;
 
 import com.google.firebase.Timestamp;
 import com.google.firebase.auth.FirebaseAuth;
@@ -105,6 +106,9 @@ public class ProfileFragment extends Fragment {
 
             db.collection("Users").document(thisUser.getUsername())
                     .set(user, SetOptions.merge());
+            Toast.makeText(thisContext,
+                    "Account information updated.",
+                    Toast.LENGTH_SHORT).show();
         });
 
         signOutbtn.setOnClickListener(view1 -> {
